+++ spec/models/repository_spec.rb
@@ -12,6 +12,7 @@ describe Repository do
     it { is_expected.to have_many(:pull_requests).dependent(:destroy) }
     it { is_expected.to have_one(:access_token).dependent(:destroy) }
     it { is_expected.to have_many(:insights).dependent(:destroy) }
+    it { is_expected.to have_many(:repository_insights).class_name('::Repositories::Insight').dependent(:destroy) }
 
     it {
       is_expected.to have_many(:pull_requests_comments).class_name('::PullRequests::Comment').through(:pull_requests)