+++ spec/services/insights/average_time/for_review_service_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 describe Insights::AverageTime::ForReviewService, type: :service do
-  subject(:service_call) { described_class.call(insightable: insightable) }
+  subject(:service_call) { described_class.call(insightable: insightable, pull_requests_ids: pull_requests_ids) }
 
   let!(:first_monday) {
     date = DateTime.now.beginning_of_month + 7.days
@@ -19,6 +19,7 @@ describe Insights::AverageTime::ForReviewService, type: :service do
   let!(:prr3) {
     create :pull_requests_review, pull_request: pr2, entity: entity2, review_created_at: first_monday + 13.hours
   }
+  let(:pull_requests_ids) { [pr1.id, pr2.id] }
 
   before do
     create :pull_requests_review, pull_request: pr1, entity: entity1, review_created_at: first_monday + 1.day + 14.hours
@@ -28,7 +29,7 @@ describe Insights::AverageTime::ForReviewService, type: :service do
     let(:insightable) { repository }
 
     it 'generates average time and succeeds', :aggregate_failures do
-      expect(service_call.result).to eq({ entity1.id => 25_200, entity2.id => 3_600 })
+      expect(service_call.result).to eq({ entity1.id => [43_200, 7_200], entity2.id => [3_600] })
       expect(service_call.success?).to be_truthy
     end
 
@@ -36,7 +37,7 @@ describe Insights::AverageTime::ForReviewService, type: :service do
       before { pr1.update!(pull_created_at: nil) }
 
       it 'generates average time without such PRs and succeeds', :aggregate_failures do
-        expect(service_call.result).to eq({ entity2.id => 3_600 })
+        expect(service_call.result).to eq({ entity2.id => [3_600] })
         expect(service_call.success?).to be_truthy
       end
     end
@@ -45,7 +46,7 @@ describe Insights::AverageTime::ForReviewService, type: :service do
       before { pr1.update!(pull_created_at: first_monday + 2.days + 1.hour) }
 
       it 'generates less average time for such PRs and succeeds', :aggregate_failures do
-        expect(service_call.result).to eq({ entity1.id => 1, entity2.id => 3_600 })
+        expect(service_call.result).to eq({ entity1.id => [1, 1], entity2.id => [3_600] })
         expect(service_call.success?).to be_truthy
       end
     end
@@ -55,7 +56,7 @@ describe Insights::AverageTime::ForReviewService, type: :service do
     let(:insightable) { repository.company }
 
     it 'generates average time and succeeds', :aggregate_failures do
-      expect(service_call.result).to eq({ entity1.id => 25_200, entity2.id => 3_600 })
+      expect(service_call.result).to eq({ entity1.id => [43_200, 7_200], entity2.id => [3_600] })
       expect(service_call.success?).to be_truthy
     end
   end
@@ -74,7 +75,7 @@ describe Insights::AverageTime::ForReviewService, type: :service do
       let(:insightable) { repository }
 
       it 'generates average time and succeeds', :aggregate_failures do
-        expect(service_call.result).to eq({ entity1.id => 12_600, entity2.id => 3_600 })
+        expect(service_call.result).to eq({ entity1.id => [18_000, 7_200], entity2.id => [3_600] })
         expect(service_call.success?).to be_truthy
       end
     end
@@ -83,7 +84,7 @@ describe Insights::AverageTime::ForReviewService, type: :service do
       let(:insightable) { repository.company }
 
       it 'generates average time and succeeds', :aggregate_failures do
-        expect(service_call.result).to eq({ entity1.id => 12_600, entity2.id => 3_600 })
+        expect(service_call.result).to eq({ entity1.id => [18_000, 7_200], entity2.id => [3_600] })
         expect(service_call.success?).to be_truthy
       end
     end
@@ -95,7 +96,7 @@ describe Insights::AverageTime::ForReviewService, type: :service do
         let(:insightable) { repository }
 
         it 'generates average time and succeeds', :aggregate_failures do
-          expect(service_call.result).to eq({ entity1.id => 28_800, entity2.id => 3_600 })
+          expect(service_call.result).to eq({ entity1.id => [50_400, 7_200], entity2.id => [3_600] })
           expect(service_call.success?).to be_truthy
         end
       end
@@ -104,7 +105,7 @@ describe Insights::AverageTime::ForReviewService, type: :service do
         let(:insightable) { repository.company }
 
         it 'generates average time and succeeds', :aggregate_failures do
-          expect(service_call.result).to eq({ entity1.id => 28_800, entity2.id => 3_600 })
+          expect(service_call.result).to eq({ entity1.id => [50_400, 7_200], entity2.id => [3_600] })
           expect(service_call.success?).to be_truthy
         end
       end
@@ -116,7 +117,7 @@ describe Insights::AverageTime::ForReviewService, type: :service do
           let(:insightable) { repository }
 
           it 'generates average time and succeeds', :aggregate_failures do
-            expect(service_call.result).to eq({ entity1.id => 28_800, entity2.id => 14_400 })
+            expect(service_call.result).to eq({ entity1.id => [50_400, 7_200], entity2.id => [14_400] })
             expect(service_call.success?).to be_truthy
           end
         end
@@ -125,7 +126,7 @@ describe Insights::AverageTime::ForReviewService, type: :service do
           let(:insightable) { repository.company }
 
           it 'generates average time and succeeds', :aggregate_failures do
-            expect(service_call.result).to eq({ entity1.id => 28_800, entity2.id => 14_400 })
+            expect(service_call.result).to eq({ entity1.id => [50_400, 7_200], entity2.id => [14_400] })
             expect(service_call.success?).to be_truthy
           end
         end
@@ -138,7 +139,7 @@ describe Insights::AverageTime::ForReviewService, type: :service do
           let(:insightable) { repository }
 
           it 'generates average time and succeeds', :aggregate_failures do
-            expect(service_call.result).to eq({ entity1.id => 28_800, entity2.id => 32_400 })
+            expect(service_call.result).to eq({ entity1.id => [50_400, 7_200], entity2.id => [32_400] })
             expect(service_call.success?).to be_truthy
           end
         end
@@ -147,7 +148,7 @@ describe Insights::AverageTime::ForReviewService, type: :service do
           let(:insightable) { repository.company }
 
           it 'generates average time and succeeds', :aggregate_failures do
-            expect(service_call.result).to eq({ entity1.id => 28_800, entity2.id => 32_400 })
+            expect(service_call.result).to eq({ entity1.id => [50_400, 7_200], entity2.id => [32_400] })
             expect(service_call.success?).to be_truthy
           end
         end
@@ -163,7 +164,7 @@ describe Insights::AverageTime::ForReviewService, type: :service do
           let(:insightable) { repository }
 
           it 'generates average time and succeeds', :aggregate_failures do
-            expect(service_call.result).to eq({ entity1.id => 28_800, entity2.id => 0 })
+            expect(service_call.result).to eq({ entity1.id => [50_400, 7_200], entity2.id => [0] })
             expect(service_call.success?).to be_truthy
           end
         end
@@ -172,7 +173,7 @@ describe Insights::AverageTime::ForReviewService, type: :service do
           let(:insightable) { repository.company }
 
           it 'generates average time and succeeds', :aggregate_failures do
-            expect(service_call.result).to eq({ entity1.id => 28_800, entity2.id => 0 })
+            expect(service_call.result).to eq({ entity1.id => [50_400, 7_200], entity2.id => [0] })
             expect(service_call.success?).to be_truthy
           end
         end