+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - container with dependencies injection for common used classes
+- repository insights
 
 ## [1.4.1] - 2023-09-15
 ### Added